backend/requirements.txt
django
pillow
djangorestframework
djangorestframework-simplejwt
django-cors-headers